+++ src/stores/elasticsearch.js
@@ -1,29 +1,6 @@
 
 import elasticsearch from 'elasticsearch';
-import BigNumber from 'bignumber.js';
-
-const serialize = (event) => {
-  const doc = Object.assign({}, event);
-  for (const key in event.args) {
-    if (event.args[key] instanceof BigNumber) {
-      doc.args[key] = {
-        type: 'BigNumber',
-        value: event.args[key].toString(),
-      };
-    }
-  }
-  return doc;
-};
-
-const unserialize = (doc) => {
-  const event = Object.assign({}, doc);
-  for (const key in doc.args) {
-    if (doc.args[key] && doc.args[key].type === 'BigNumber') {
-      event.args[key] = new BigNumber(doc.args[key].value);
-    }
-  }
-  return event;
-};
+import { serialize, unserialize } from '../utils';
 
 const createElasticsearchIndex = async (client, eventType, config) => {
   const index = eventType.toLowerCase();