+++ src/stores/file.js
@@ -2,30 +2,7 @@
 import fs from 'fs';
 import path from 'path';
 import rimraf from 'rimraf';
-import BigNumber from 'bignumber.js';
-
-const serialize = (event) => {
-  const doc = Object.assign({}, event);
-  for (const key in event.args) {
-    if (event.args[key] instanceof BigNumber) {
-      doc.args[key] = {
-        type: 'BigNumber',
-        value: event.args[key].toString(),
-      };
-    }
-  }
-  return doc;
-};
-
-const unserialize = (doc) => {
-  const event = Object.assign({}, doc);
-  for (const key in doc.args) {
-    if (doc.args[key] && doc.args[key].type === 'BigNumber') {
-      event.args[key] = new BigNumber(doc.args[key].value);
-    }
-  }
-  return event;
-};
+import { serialize, unserialize } from '../utils';
 
 export default class FileStore {
   constructor(indexing, dbPath) {
@@ -69,6 +46,9 @@ export default class FileStore {
   get(eventType, indexId, value) {
     const indexKey = `${eventType}-${indexId}-${value}`;
     const filePath = `${this.dbPath}/${indexKey}.jsons`;
+    if (!fs.existsSync(filePath)) {
+      return [];
+    }
     return fs.readFileSync(filePath)
       .toString().split('\n')
       .filter(line => line.length > 0)